Migrate from tslint to eslint
tslint is on the way to become deprecated, and will be made completely
deprecated by the end of the year (see
https://github.com/palantir/tslint/issues/4534 for details).

This PR migrates the project from tslint to eslint, which is the
de-facto standard for linting typescript.
It switches off some rules and takes the current code as the de-facto
coding standard. Follow up work may want to tweak these rules (and the
code accordingly).